requirements.txt
Flask==3.0.0
Flask-RESTX==1.3.0
Flask-CORS==4.0.0
Flask-SocketIO==5.3.6
python-dotenv==1.0.0
GitPython==3.1.40
tree-sitter
tree-sitter-python
tree-sitter-javascript
networkx==3.2.1
chromadb==0.4.21
langchain==0.1.0
langgraph==0.0.15
huggingface-hub==0.20.2
groq==0.4.1
sentence-transformers==2.2.2
gunicorn==21.2.0
